# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tennis prediction engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fixtures/upcoming:
    get:
      operationId: getUpcomingFixtures
      tags: [matchups]
      summary: Upcoming scheduled matches
      description: >-
        Returns a rolling now-forward window of matches with a start time at or after the current
        instant, sorted soonest-first, regardless of which calendar day (UTC) they fall on. The
        window auto-extends further out when the near-term days are sparse, so the result is
        capped by `limit` rather than by a fixed clock/day boundary. Use `offset` together with
        `hasMore` in the response to page further into the window on busy days (e.g. Challenger/ITF
        days with 50+ matches before noon).
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of fixtures to return. Defaults to 50.
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          required: false
          description: Number of soonest-first fixtures to skip before returning `limit` more. Defaults to 0.
          schema:
            type: integer
            minimum: 0
        - name: force
          in: query
          required: false
          description: >-
            When true, bypasses the provider's short in-memory fixtures cache for this call, so a
            user-initiated refresh can actually pull fresh data instead of silently re-serving the
            same cached response. Defaults to false (normal cached reads).
          schema:
            type: boolean
      responses:
        "200":
          description: Upcoming fixtures
          content:
            application/json:
              schema:
                type: object
                required: [fixtures, hasMore]
                properties:
                  fixtures:
                    type: array
                    items:
                      $ref: "#/components/schemas/Fixture"
                  hasMore:
                    type: boolean
                    description: True when at least one more fixture exists beyond `offset + limit` within the lookahead window.
        "502":
          description: Upstream data provider unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderError"
...
components:
  schemas:
...
    ProviderError:
      type: object
      required: [error]
      properties:
        error:
          type: string
        detail:
          type: ["string", "null"]
...
    Surface:
      type: string
      enum: [Hard, Clay, Grass, IndoorHard]
    TournamentLevel:
      type: string
      enum: [GrandSlam, Masters1000, ATP500, ATP250, WTA1000, WTA500, WTA250, Challenger, ITF, Other]
    MatchFormat:
      type: string
      enum: [BestOf3, BestOf5]
...
    Fixture:
      type: object
      required: [id, date, timeConfirmed, isLive, player1Id, player1Name, player2Id, player2Name, surface]
      properties:
        id:
          type: string
        date:
          type: string
          format: date
          description: Calendar date the match is scheduled for (YYYY-MM-DD). Always present.
        scheduledStart:
          oneOf:
            - type: "null"
            - type: string
              format: date-time
          description: >-
            Full real start instant (UTC) for this exact fixture, combining the provider's
            per-match date and time. Null when the provider did not supply a verified time for
            this fixture -- never fabricated or shared with another match. Clients must show
            "Time TBD" when null. IMPORTANT -- null must be listed first in this oneOf union.
            The generated zod.coerce.date() silently accepts a null input as epoch 1970 instead
            of failing, so if the date branch were checked first a real null would be miscoerced
            into a fake timestamp.
        timeConfirmed:
          type: boolean
          description: True only when scheduledStart reflects a real provider-supplied time for this exact fixture.
        isLive:
          type: boolean
          description: >-
            True when this fixture's confirmed scheduledStart is already in the past and the
            provider has not yet reported a winner -- i.e. the match is currently in progress.
            False for a genuinely upcoming (not yet started) fixture, and also false for an
            unconfirmed ("Time TBD") fixture, since we have no real evidence it has started.
        tournamentName:
          type: ["string", "null"]
        tournamentLevel:
          oneOf:
            - $ref: "#/components/schemas/TournamentLevel"
            - type: "null"
        round:
          type: ["string", "null"]
        surface:
          oneOf:
            - $ref: "#/components/schemas/Surface"
            - type: "null"
        indoor:
          type: ["boolean", "null"]
        matchFormat:
          $ref: "#/components/schemas/MatchFormat"
        player1Id:
          type: string
        player1Name:
          type: string
        player2Id:
          type: string
        player2Name:
          type: string
...